+++ my-app/src/components/CardForm/CvcSection/CvcSection.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import CommonSection from '../../common/CommonSection/CommonSection';
-import NumberInput from '../../common/NumberInput/NumberInput';
-import { getCardCvcError } from '../../utils/validation';
+import CommonSection from '../../../common/CommonSection/CommonSection';
+import NumberInput from '../../../common/NumberInput/NumberInput';
+import { getCardCvcError } from '../../../utils/validation';
 
 interface Props {
   value: string;
