+++ my-app/src/components/CardForm/ExpirationDateSection/ExpirationDateSection.stories.tsx
@@ -15,6 +15,6 @@ export const Second: Story = {
   render: () => {
     const [value, setValue] = useState({ month: '', year: '' });
 
-    return <ExpirationDateSection value={value} setValue={setValue} />;
+    return <ExpirationDateSection value={value} updateValue={setValue} />;
   },
 };
