+++ my-app/src/components/CardForm/ExpirationDateSection/ExpirationDateSection.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from 'react';
-import CommonSection from '../../common/CommonSection/CommonSection';
-import NumberInput from '../../common/NumberInput/NumberInput';
-import { getCardMonthError, getCardYearError } from '../../utils/validation';
+import { getCardMonthError, getCardYearError } from '../../../utils/validation';
+import CommonSection from '../../../common/CommonSection/CommonSection';
+import NumberInput from '../../../common/NumberInput/NumberInput';
 
 interface Props {
   value: {
