+++ my-app/src/components/CardForm/SubmitButtonSection/SubmitButtonSection.stories.tsx
@@ -11,20 +11,24 @@ type Story = StoryObj<typeof SubmitButtonSection>;
 
 export const Enabled: Story = {
   args: {
-    cardNumber: '12312312',
-    cardExpirationDate: { month: '12', year: '30' },
-    cardCvc: '123',
-    cardIssuer: '신한카드',
-    cardPassword: '12',
+    cardForm: {
+      cardNumber: '1234123412341234',
+      cardExpirationDate: { month: '12', year: '30' },
+      cardCvc: '123',
+      cardIssuer: '신한카드',
+      cardPassword: '12',
+    },
   },
 };
 
 export const Disabled: Story = {
   args: {
-    cardNumber: '1234123',
-    cardExpirationDate: { month: '12', year: '30' },
-    cardCvc: '123',
-    cardIssuer: '신한카드',
-    cardPassword: '12',
+    cardForm: {
+      cardNumber: '1234123',
+      cardExpirationDate: { month: '12', year: '30' },
+      cardCvc: '123',
+      cardIssuer: '신한카드',
+      cardPassword: '12',
+    },
   },
 };
